+++ src/components/SimpleNormalizedChart/index.ts
@@ -1,0 +1,4 @@
+export {SimpleNormalizedChart} from './SimpleNormalizedChart';
+export {BarSegment} from './components/BarSegment';
+export type {SimpleNormalizedChartProps} from './SimpleNormalizedChart';
+export type {Size} from './types';